test(react): remove storybook system axis demos
