+++ packages/react/src/components/ui/dropdown-menu/DropdownMenu.stories.tsx
@@ -526,73 +526,6 @@ export const AllVariants: Story = {
   },
 };
 
-export const ItemPaddingPinnedAcrossModes: Story = {
-  parameters: {
-    a11y: { test: 'off' },
-    docs: {
-      description: {
-        story:
-          'Regression sentinel: opened `DropdownMenuItem` padding stays constant across document spacing modes.',
-      },
-    },
-  },
-  render: () => (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="outline">Open Menu</Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent>
-        <DropdownMenuItem>Profile</DropdownMenuItem>
-        <DropdownMenuItem>Settings</DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  ),
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const trigger = canvas.getByRole('button', { name: 'Open Menu' });
-    const originalMode = document.documentElement.getAttribute('data-style');
-
-    try {
-      for (const mode of ['compact', 'spacious'] as const) {
-        document.documentElement.setAttribute('data-style', mode);
-        await userEvent.click(trigger);
-
-        const item = await waitFor(() => {
-          const el = document.querySelector<HTMLElement>(
-            '[data-slot="dropdown-menu-item"]'
-          );
-          if (!el) throw new Error('dropdown item not visible yet');
-          return el;
-        });
-
-        const styles = getComputedStyle(item);
-        expect(styles.paddingTop).toBe('6px');
-        expect(styles.paddingBottom).toBe('6px');
-
-        await userEvent.keyboard('{Escape}');
-        await waitFor(() => {
-          expect(
-            document.querySelector('[data-slot="dropdown-menu-item"]')
-          ).toBeNull();
-        });
-      }
-    } finally {
-      if (document.querySelector('[role="menu"]')) {
-        await userEvent.keyboard('{Escape}');
-        await waitFor(() => {
-          expect(document.querySelector('[role="menu"]')).toBeNull();
-        });
-      }
-
-      if (originalMode === null) {
-        document.documentElement.removeAttribute('data-style');
-      } else {
-        document.documentElement.setAttribute('data-style', originalMode);
-      }
-    }
-  },
-};
-
 // ============================================
 // A11Y is tested automatically on ALL stories
 // via addon-a11y with test: 'error'
